refactor: split job and worker
